Update dataset_creator.sh
Include the gyros/ folder in training and test data.

Since gyros has only 300 images, and the rest of the
categories have around 1000 images, I decided to
have almost the same number of images for every
category. 

Therefore I include the first 200 images of each 
category in the training set and the next 100 to the 
test set. 

The rest images are ignored.